implement TokenService

+++ src/main/kotlin/com/jorgereyesdev/spring_security/domain/models/User.kt
@@ -3,7 +3,7 @@ package com.jorgereyesdev.spring_security.domain.models
 import com.jorgereyesdev.spring_security.infrastructure.entities.TokenEntity
 
 data class User(
-    var id: Int? = null,
+    var id: Long? = null,
     var username: String,
     var password: String,
     var enable: Boolean,
